US90806 d2l-search-widget update + switch to d2l-fetch
The recent version bump to d2l-search-widget removed the cache-responses functionality, so remove it. Instead, caching should be done by using the d2l-fetch-simple-cache middleware (d2l-search-widget now uses d2l-fetch internally).

We noticed some significant performance gains by switching from d2l-ajax to d2l-fetch in d2l-my-courses; not really expecting to see as much of a benefit here, but the API is nicer to work with, and it allows us to add caching and some other goodness page-wide, which is legit.